Various fixes for various compilers...
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4424 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/html/htmlpars.cpp
@@ -81,7 +81,7 @@ void wxHtmlParser::DoParsing(int begin_pos, int end_pos)
     i = begin_pos;
 
     while (i < end_pos) {
-        c = m_Source[i];
+        c = m_Source[(unsigned int) i];
 
         // continue building word:
         if (c != '<') {
