+++ src/html/m_layout.cpp
@@ -148,7 +148,7 @@ TAG_HANDLER_BEGIN(TITLE, "TITLE")
                 wxString title = "";
                 wxString *src = m_WParser -> GetSource();
 
-                for (int i = tag.GetBeginPos(); i < tag.GetEndPos1(); i++) title += (*src)[i];
+                for (int i = tag.GetBeginPos(); i < tag.GetEndPos1(); i++) title += (*src)[(unsigned int) i];
                 wfr -> SetTitle(title);
             }
         }
